Move continuous space modules to experimental
