+++ docs/api/geometry.md
@@ -1,12 +1,13 @@
+
 (netket_geometry_api)=
-# netket.geometry
+# netket.experimental.geometry
 
 ```{eval-rst}
-.. currentmodule:: netket.geometry
+.. currentmodule:: netket.experimental.geometry
 ```
 
-The :mod:`netket.geometry` module contains helper classes to describe the geometry of continuous spaces.
-These objects can be passed to {class}`~netket.hilbert.Particle` and provide utilities such as distance computations.
+The :mod:`netket.experimental.geometry` module contains helper classes to describe the geometry of continuous spaces.
+These objects can be passed to {class}`~netket.experimental.hilbert.Particle` and provide utilities such as distance computations.
 
 ```{eval-rst}
 .. autosummary::
