+++ netket/hilbert/__init__.py
@@ -19,13 +19,11 @@ from .abstract_hilbert import AbstractHilbert
 from .discrete_hilbert import DiscreteHilbert
 from .homogeneous import HomogeneousHilbert
 
-from .continuous_hilbert import ContinuousHilbert
 
 from .doubled_hilbert import DoubledHilbert
 from .spin import Spin
 from .fock import Fock
 from .qubit import Qubit
-from .particle import Particle
 from .spin_orbital_fermions import SpinOrbitalFermions
 
 from .tensor_hilbert import TensorHilbert
