+++ test/models/test_deepset.py
@@ -58,7 +58,9 @@ def test_deepset_model_output():
 )
 def test_rel_dist_deepsets(cusp_exponent, L):
     d = len(L) if hasattr(L, "__len__") else 1
-    hilb = nk.hilbert.Particle(N=2, geometry=nk.geometry.Cell(d=d, L=L, pbc=True))
+    hilb = nk.experimental.hilbert.Particle(
+        N=2, geometry=nk.experimental.geometry.Cell(d=d, L=L, pbc=True)
+    )
     sdim = len(hilb.domain)
     x = jnp.hstack([jnp.ones(4), -jnp.ones(4)]).reshape(1, -1)
     xp = jnp.roll(x, sdim)
@@ -76,7 +78,9 @@ def test_rel_dist_deepsets(cusp_exponent, L):
 
 
 def test_rel_dist_deepsets_error():
-    hilb = nk.hilbert.Particle(N=2, geometry=nk.geometry.Cell(d=1, L=1.0, pbc=True))
+    hilb = nk.experimental.hilbert.Particle(
+        N=2, geometry=nk.experimental.geometry.Cell(d=1, L=1.0, pbc=True)
+    )
     sdim = len(hilb.domain)
 
     x = jnp.hstack([jnp.ones(4), -jnp.ones(4)]).reshape(1, -1)
@@ -104,9 +108,9 @@ def test_rel_dist_deepsets_error():
 
     with pytest.raises(ValueError):
         ds = nk.models.DeepSetRelDistance(
-            hilbert=nk.hilbert.Particle(
+            hilbert=nk.experimental.hilbert.Particle(
                 N=2,
-                geometry=nk.geometry.Cell(d=1, L=1.0, pbc=False),
+                geometry=nk.experimental.geometry.Cell(d=1, L=1.0, pbc=False),
             ),
             layers_phi=2,
             layers_rho=2,
