+++ test/nn/test_blocks.py
@@ -105,8 +105,8 @@ def test_deepset():
     """Test the permutation invariance"""
     L = (1.0, 1.0)
     n_particles = 6
-    hilb = nk.hilbert.Particle(
-        N=n_particles, geometry=nk.geometry.Cell(d=len(L), L=L, pbc=True)
+    hilb = nk.experimental.hilbert.Particle(
+        N=n_particles, geometry=nk.experimental.geometry.Cell(d=len(L), L=L, pbc=True)
     )
     sdim = len(hilb.domain)
     key = jax.random.PRNGKey(42)
