+++ test/operator/test_continuous_operator.py
@@ -19,9 +19,11 @@ def v2(x):
 v2_vec = jax.vmap(v2)
 
 
-hilb = netket.hilbert.Particle(N=1, geometry=netket.geometry.FreeSpace(d=1))
-hilb2 = netket.hilbert.Particle(
-    N=2, geometry=netket.geometry.Cell(d=1, L=5.0, pbc=True)
+hilb = netket.experimental.hilbert.Particle(
+    N=1, geometry=netket.experimental.geometry.FreeSpace(d=1)
+)
+hilb2 = netket.experimental.hilbert.Particle(
+    N=2, geometry=netket.experimental.geometry.Cell(d=1, L=5.0, pbc=True)
 )
 
 # potential operators
